Initial model setup and small test setup. -----------------------------------------
With the create test, a database was made in Atlas Cloud named "test" in which the created user is made. Currently, any time the tests run a new user will be created in a collection called "users" each with the same username and password but a different ObjectId.

Currently the schema's follow the pattern shown in Trello under the card "MongoDB: Database opzetten".