Safe version, still contains bug in removal. ---------------------------------
When removing a thread, all the underlying comments should be removed however the top layer of comments seems to stay with the populated comments nested within.
With the removal test turned off all (three) comments are in the database with it's children. With it turned on, only the actual comment on the removed thread exists and has children but the children of that comment are gone.